requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch
torchvision
transformers
matplotlib
